+++ tests/Unit/Validators/ArbitrarySizeValidatorTest.php
@@ -7,6 +7,8 @@ test('is arbitrary size', function ($input, $output) {
 })->with([
     ['[size:2px]', true],
     ['[size:bla]', true],
+    ['[length:bla]', true],
+    ['[percentage:bla]', true],
 
     ['[2px]', false],
     ['[bla]', false],